Move vaccount's require_auth to utility function
This function will go in a shared header later so it can be imported by other
contracts.

Some strange behaviour occurs when `require_sig` is called from within
`require_auth`. There are frequent `iterator out of bounds` errors in the smart
contract thrown by the `require_sig` and `pub2addr` functions. The cause of this
is the std::get<...> calls on the variants. In the end making `require_sig` an
inline function fixes the issue, but I honestly don't know why this is
necessary.