Do not allow signatures on EOS virtual transactions
previously a signature could be added on an EOS transaction, wich was ignored by
the smart contract. this can be abused because someone could use the relayer
account as the from_account.

the relayer should only work for EOS accounts. by not allowing sig on EOS
transactions in the smart contract we will avoid mistakes by relayers.