Simplify payout logic in force contract
the payout action contained a lot of logic on which payouts should be paid out
for an account ID. it was looping accross all payments to check which are
elegible for payout and batching them into a single transaction. this would be a
bit more effcicient transaction wise, but it's prone to bugs (e.g. it would
already fail if there are payments with different symbols or tokens).

let's move this "smart" logic from the contract to the calling code (SDK)
instead.