ci: Compare remote vs configured code during deployment
if a code hash is configured and it equals the remote code, then the
local contract will not be deployed. this allows us to diverge mainnet
deployments from code changes on the main branch.

when we are ready to deploy a mainnet contract, the contract hash
should be commited to this file for it to deploy