force: Store `batch_id` as `uint32` in submission and reservation
before we were storing a uint64 (concatenation of campaign_id and
batch_idx). but as campaign_id is already present, we can save space
by just storing batch_idx. it is also more convenient for clients, as
there is no need to disect the batch_id in order to find the batch
index.